nexus911/requirements.txt
# Nexus911 — Multi-Agent Emergency Coordination
# Core
google-adk>=1.2.0
google-genai>=1.10.0
google-cloud-firestore>=2.19.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=14.0
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.0.0

# Utilities
httpx>=0.27.0
python-multipart>=0.0.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
